requirements.txt
requests
openpyxl
pandas
pytz
datefinder
mongoengine
dnspython
tzlocal
genshin
auraxium
deprecated
pydantic<2
discord.py
fireo